Utilizes a 2 bytes+ for result in alu ops
this is because we will lose data and won't likely have a maxed out result otherwise when doing operations that result in 256+

+++ ls8/alu.c
@@ -8,7 +8,7 @@
  * ALU
  */
 void alu(struct cpu *cpu, enum alu_op op, byte regA, byte regB) {
-  byte result;
+  short int result;
   switch (op) {
     case ALU_CMP:
       if (cpu->registers[regA] == cpu->registers[regB])
